pyproject.toml
[project]
name = "illama-index-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "faiss-cpu>=1.8.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-huggingface>=0.3.1",
    "livekit-agents>=0.9.0",
    "livekit-plugins-cartesia>=0.9.0",
    "livekit-plugins-deepgram>=0.9.0",
    "livekit-plugins-google>=0.9.0",
    "livekit-plugins-langchain>=0.9.0",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-silero>=1.2.5",
    "livekit-plugins-turn-detector>=0.9.0",
    "llama-index>=0.13.1",
    "llama-index-embeddings-huggingface>=0.6.0",
    "llama-index-llms-google-genai>=0.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.1.0",
    "typing-extensions>=4.8.0",
]
